Pull request #4191: kaizen: remove redundant semicolum and add explicit cast
Merge in SNORT/snort3 from ~YVELYKOZ/snort3:fix_cppchecks_kaizen to master

Squashed commit of the following:

commit a7e95f4068b1a38923be160c74aeac5e1e19298d
Author: Yehor Velykozhon <yvelykoz@cisco.com>
Date:   Thu Feb 1 16:12:57 2024 +0200

    kaizen: remove redundant semicolon and add explicit cast

+++ src/network_inspectors/kaizen/kaizen_inspector.cc
@@ -93,7 +93,7 @@ void HttpBodyHandler::handle(DataEvent& de, Flow*)
     kaizen_stats.client_body_bytes += body_len;
 
     debug_logf(kaizen_trace, TRACE_CLASSIFIER, nullptr, "input (body): %.*s\n", body_len, body);
-    debug_logf(kaizen_trace, TRACE_CLASSIFIER, nullptr, "output: %f\n", output);
+    debug_logf(kaizen_trace, TRACE_CLASSIFIER, nullptr, "output: %f\n", static_cast<double>(output));
 
     if ((double)output > config.http_param_threshold)
     {
@@ -148,7 +148,7 @@ void HttpUriHandler::handle(DataEvent& de, Flow*)
     kaizen_stats.uri_bytes += query_len;
 
     debug_logf(kaizen_trace, TRACE_CLASSIFIER, nullptr, "input (query): %.*s\n", query_len, query);
-    debug_logf(kaizen_trace, TRACE_CLASSIFIER, nullptr, "output: %f\n", output);
+    debug_logf(kaizen_trace, TRACE_CLASSIFIER, nullptr, "output: %f\n", static_cast<double>(output));
 
     if ((double)output > config.http_param_threshold)
     {
